requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
apimatic-core==0.2.8
apimatic-core-interfaces==0.1.4
apimatic-requests-client-adapter==0.1.6
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
black==23.12.1
boto3==1.34.77
botocore==1.34.77
Brotli==1.1.0
CacheControl==0.12.14
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==41.0.7
cssselect2==0.7.0
defusedxml==0.8.0rc2
deprecation==2.1.0
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.8
django-celery-beat==2.6.0
django-ckeditor-5==0.2.12
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-filter==23.5
django-js-asset==2.2.0
django-nine==0.2.7
django-rest-framework==0.1.0
django-storages==1.14.2
django-templated-mail==1.1.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
easypost==9.1.0
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
enum34==1.1.10
exceptiongroup==1.2.1
fonttools==4.47.2
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jmespath==1.0.1
jsonpickle==3.0.3
jsonpointer==2.4
kombu==5.3.5
msgpack==1.0.8
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.7.1
pydantic-core==2.18.2
pydyf==0.8.0
PyJWT==2.8.0
pyphen==0.14.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
pyzipcode==3.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sorl-thumbnail==12.10.0
sqlparse==0.4.4
squareup==35.1.0.20240320
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.66.2
twilio==8.11.1
typing-extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
weasyprint==60.2
webencodings==0.5.1
yarl==1.9.4
zopfli==0.2.3
